+++ RELEASE-NOTES.md
@@ -30,6 +30,7 @@
   * These scripts will execute whenever a Native App Package is created or updated.
   * Currently only supports SQL scripts: `post_deploy: [{sql_script: script.sql}]`
 * Added `snow spcs service execute-job` command, which supports creating and executing a job service in the current schema.
+* Added `snow connection remove` command, which removes a connection from the configuration file.
 
 ## Fixes and improvements
 * Fixed problem with whitespaces in `snow connection add` command